minor: switch arg type `max-upload` to `uint64`

+++ main.go
@@ -107,7 +107,7 @@ func main() {
 	args.String("queue", "q", "", "queue file (default: <out>/queue.json)").Env("QUEUE_FILE")
 	args.String("ffmpeg", "", "ffmpeg", "path to the ffmpeg binary").Env("FFMPEG_PATH")
 	args.String("ffprobe", "", "ffprobe", "path to the ffprobe binary").Env("FFPROBE_PATH")
-	args.Int("max-upload", "", 16<<30, "largest accepted upload in bytes").Env("MAX_UPLOAD_SIZE")
+	args.Uint64("max-upload", "", 16<<30, "largest accepted upload in bytes").Env("MAX_UPLOAD_SIZE")
 	args.Bool("allow-commands", "", false, "let the post-queue action run a shell command").Env("ALLOW_COMMAND")
 	must := func(err error) {
 		if err != nil {
@@ -129,7 +129,7 @@ func main() {
 	queueFile := cfg.String("queue")
 	ffmpegBin := cfg.String("ffmpeg")
 	ffprobeBin := cfg.String("ffprobe")
-	maxUpload := cfg.Int("max-upload")
+	maxUpload := cfg.Uint64("max-upload")
 	allowCmds := cfg.Bool("allow-commands")
 
 	mediaRoot, err := filepath.Abs(root)
@@ -177,7 +177,8 @@ func main() {
 		broker: broker, jobs: manager, store: store, presets: presetStore, frames: newFrameCache(256 << 20),
 		log:       logger.Named("web"),
 		monitor:   newSystemMonitor(),
-		maxUpload: int64(maxUpload), allowCmds: allowCmds,
+		maxUpload: int64(maxUpload),
+		allowCmds: allowCmds,
 	}
 
 	httpSrv := &http.Server{
